backend/pyproject.toml
[project]
name = "litestar-keycloak"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Joren Six", email = "joren.six@ugent.be"},
]
dependencies = [
    "python-keycloak>=3.7.0",
    "python-decouple>=3.8",
    "litestar[standard]>=2.5.3",
    "uvicorn>=0.27.0.post1",
    "requests>=2.31.0",
    "python-jose[cryptography]>=3.3.0",
]
requires-python= ">= 3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
